app/requirements.txt
psycopg2==2.9.9
psycopg2-binary==2.9.9
elasticsearch==8.15.1
python-dotenv==1.0.1
streamlit==1.39.0
boto3
sentence-transformers